Move everything concerning overlay to seperate folder, to create more structure
