+++ components/datagrid/HeaderRowOverlay/BulkActions.js
@@ -1,6 +1,6 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
-import Box from '../box';
+import Box from '../../box';
 import theme from './theme.css';
 
 class BulkActions extends PureComponent {
