+++ components/datagrid/HeaderRowOverlay/theme.css
@@ -1,0 +1,15 @@
+@import '@teamleader/ui-colors';
+@import '@teamleader/ui-typography';
+
+.header-row-overlap {
+  background: var(--color-neutral-light);
+  position: absolute;
+  left: 48px;
+  right: 0;
+  height: 48px;
+  z-index: 3;
+}
+
+.num-selected-rows-number {
+  font-family: var(--font-family-medium);
+}
