+++ components/datagrid/theme.css
@@ -220,17 +220,3 @@
   justify-content: flex-end;
   text-align: right;
 }
-
-/* Header row overlap */
-.header-row-overlap {
-  background: var(--color-neutral-light);
-  position: absolute;
-  left: 48px;
-  right: 0;
-  height: 48px;
-  z-index: 3;
-}
-
-.num-selected-rows-number {
-  font-family: var(--font-family-medium);
-}
